refactor: split app helper modules
